[REFACTOR] apply first review
