+++ src/main/kotlin/com/pravo/pravo/domain/member/dto/MemberPointLogResponseDTO.java
@@ -7,7 +7,7 @@ public record MemberPointLogResponseDTO(
     @Schema(description = "약속 이름", example = "약속 이름")
     String promiseName,
 
-    @Schema(description = "pointLogStatus", example = "PLUS(적립+)/MINUS(사용-)/FINE(별금-)")
+    @Schema(description = "pointLogStatus", example = "PLUS(적립+)/MINUS(사용-)")
     String pointLogStatus,
 
     @Schema(description = "결제 금액", example = "10000")
