+++ src/main/kotlin/com/pravo/pravo/domain/member/service/MemberService.java
@@ -1,10 +1,21 @@
 package com.pravo.pravo.domain.member.service;
 
+import com.pravo.pravo.domain.fine.model.FineLog;
+import com.pravo.pravo.domain.fine.repository.FineLogRepository;
 import com.pravo.pravo.domain.member.dto.LoginRequestDTO;
 import com.pravo.pravo.domain.member.dto.LoginResponseDTO;
+import com.pravo.pravo.domain.member.dto.MemberPaymentLogResponseDTO;
+import com.pravo.pravo.domain.member.dto.MemberPointLogResponseDTO;
 import com.pravo.pravo.domain.member.dto.MyPageResponseDTO;
 import com.pravo.pravo.domain.member.model.Member;
 import com.pravo.pravo.domain.member.repository.MemberRepository;
+import com.pravo.pravo.domain.payment.enums.PaymentStatus;
+import com.pravo.pravo.domain.payment.model.PaymentLog;
+import com.pravo.pravo.domain.payment.repository.PaymentLogRepository;
+import com.pravo.pravo.domain.point.model.PointLog;
+import com.pravo.pravo.domain.point.repository.PointLogRepository;
+import com.pravo.pravo.domain.promise.model.Promise;
+import com.pravo.pravo.domain.promise.repository.PromiseRepository;
 import com.pravo.pravo.global.error.ErrorCode;
 import com.pravo.pravo.global.error.exception.BaseException;
 import com.pravo.pravo.global.error.exception.NotFoundException;
@@ -14,7 +25,11 @@ import com.pravo.pravo.global.jwt.JwtTokenProvider;
 import com.pravo.pravo.global.jwt.JwtTokens;
 import com.pravo.pravo.global.jwt.JwtTokensGenerator;
 import jakarta.persistence.EntityNotFoundException;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
@@ -31,14 +46,25 @@ public class MemberService {
     private final RedisTemplate<String, String> redisTemplate;
     private final S3Service s3Service;
 
+    private final PromiseRepository promiseRepository;
+    private final PaymentLogRepository paymentLogRepository;
+    private final PointLogRepository pointLogRepository;
+    private final FineLogRepository fineLogRepository;
+
     public MemberService(MemberRepository memberRepository, JwtTokensGenerator jwtTokensGenerator,
         RedisTemplate<String, String> redisTemplate, JwtTokenProvider jwtTokenProvider,
-        S3Service s3Service) {
+        S3Service s3Service, PromiseRepository promiseRepository,
+        PaymentLogRepository paymentLogRepository, PointLogRepository pointLogRepository,
+        FineLogRepository fineLogRepository) {
         this.memberRepository = memberRepository;
         this.jwtTokensGenerator = jwtTokensGenerator;
         this.redisTemplate = redisTemplate;
         this.jwtTokenProvider = jwtTokenProvider;
         this.s3Service = s3Service;
+        this.promiseRepository = promiseRepository;
+        this.paymentLogRepository = paymentLogRepository;
+        this.pointLogRepository = pointLogRepository;
+        this.fineLogRepository = fineLogRepository;
     }
 
     public LoginResponseDTO login(LoginRequestDTO loginRequestDto) {
@@ -135,5 +161,64 @@ public class MemberService {
     public Member getMemberById(Long memberId) {
         return memberRepository.findById(memberId)
             .orElseThrow(() -> new EntityNotFoundException("멤버를 찾을 수 없습니다"));
+    }
+
+    public List<MemberPaymentLogResponseDTO> getMemberPaymentLog(Long memberId) {
+        List<PaymentLog> paymentLogs = paymentLogRepository.findByMemberIdAndPaymentStatus(
+            memberId,
+            List.of(PaymentStatus.COMPLETED, PaymentStatus.CANCELED));
+
+        return paymentLogs.stream()
+            .map(paymentLog -> new MemberPaymentLogResponseDTO(
+                paymentLog.getPromise().getName(),
+                paymentLog.getApprovedAt().toString(),
+                paymentLog.getBalanceAmount(),
+                paymentLog.getPaymentStatus().name()
+            ))
+            .collect(Collectors.toList());
+    }
+
+    public List<MemberPointLogResponseDTO> getMemberPointLog(Long memberId) {
+        List<PointLog> pointLogs = pointLogRepository.findByMemberId(memberId);
+        List<FineLog> fineLogs = fineLogRepository.findByMemberId(memberId);
+
+        List<MemberPointLogResponseDTO> combinedLogs = new ArrayList<>();
+
+        // pointlog
+        combinedLogs.addAll(
+            pointLogs.stream()
+                .map(pointLog -> {
+                    Promise promise = promiseRepository.findById(pointLog.getPromiseId())
+                        .orElseThrow(
+                            () -> new NotFoundException(ErrorCode.NOT_FOUND, "약속을 찾을 수 없습니다"));
+                    return new MemberPointLogResponseDTO(
+                        promise.getName(),
+                        pointLog.getPointLogStatus().toString(),
+                        pointLog.getAmount(),
+                        pointLog.getUpdatedAt()
+                    );
+                })
+                .toList()
+        );
+
+        // finelog
+        combinedLogs.addAll(
+            fineLogs.stream()
+                .map(fineLog -> {
+                    Promise promise = promiseRepository.findById(fineLog.getPromiseId())
+                        .orElseThrow(
+                            () -> new NotFoundException(ErrorCode.NOT_FOUND, "약속을 찾을 수 없습니다"));
+                    return new MemberPointLogResponseDTO(
+                        promise.getName(),
+                        "FINE",
+                        fineLog.getAmount(),
+                        fineLog.getUpdatedAt()
+                    );
+                })
+                .toList()
+        );
+
+        combinedLogs.sort(Comparator.comparing(MemberPointLogResponseDTO::pointDate).reversed());
+        return combinedLogs;
     }
 }
