+++ src/main/kotlin/com/pravo/pravo/domain/payment/repository/PaymentLogRepository.java
@@ -5,10 +5,18 @@ import com.pravo.pravo.domain.payment.model.PaymentLog;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 public interface PaymentLogRepository extends JpaRepository<PaymentLog, String> {
 
     Optional<PaymentLog> findByMemberIdAndPromiseId(Long memberId, Long promiseId);
 
-    List<PaymentLog> findByMemberIdAndPaymentStatus(Long memberId, PaymentStatus paymentStatus);
+    @Query("""
+            SELECT pl FROM PaymentLog pl
+            JOIN FETCH pl.promise p
+            WHERE pl.memberId = :memberId
+            AND pl.paymentStatus IN :paymentStatuses
+        """)
+    List<PaymentLog> findByMemberIdAndPaymentStatus(Long memberId,
+        List<PaymentStatus> paymentStatuses);
 }
